+++ tests/report_generators.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import datetime
 
 from django.db.models import Sum, Count
@@ -34,7 +36,7 @@ class GeneratorWithAttrAsColumn(GenericGenerator):
     columns = ["get_data", "slug", "name"]
 
     def get_data(self, obj):
-        return ""
+        return obj["name"]
 
     get_data.verbose_name = "My Verbose Name"
 
@@ -106,9 +108,14 @@ class CrosstabOnTraversingField(ReportGenerator):
 
 class ClientTotalBalance(ReportGenerator):
     report_model = SimpleSales
-    date_field = "doc_date"
+    # date_field = "doc_date"
     group_by = "client"
-    columns = ["slug", "name", "__balance__", "__total__"]
+    columns = [
+        "slug",
+        "name",
+        "__balance__",
+        ComputationField.create(Sum, "value", name="__total__", verbose_name=_("Sales")),
+    ]
 
 
 class TotalBalanceWithQueryset(ReportGenerator):
@@ -166,7 +173,12 @@ class ProductTotalSales(ReportGenerator):
         "average_value",
     ]
 
-    def get_object_sku(self, obj, data):
+    def get_object_sku(self, obj: dict, row: dict) -> any:
+        """
+        :param obj: obj is the current row of the grouped by model , or the current row of the queryset
+        :param row: the current report row values in a dictionary
+        :return:
+        """
         return obj["sku"].upper()
 
     get_object_sku.verbose_name = "SKU ALL CAPS"